feat(site): gate user dropdown AI spend and share the usage bar
Gate the dropdown AI spend behind aibridge plus the ai-gateway-cost-control
experiment, make the bar taller, and extract a shared UsageBar reused by the
agents usage indicator.
